revert wallet creation file splitting
