+++ src/components/wallet/WalletImportDialog.vue
@@ -20,20 +20,88 @@
     v-model="show"
     max-width="680px"
   >
-    <WalletImport />
+    <v-card>
+      <v-card-title primary-title>
+        <h3 class="headline mb-3">
+          Import a wallet from private key
+        </h3>
+        <p class="mb-0 clearfix">
+          This wallet will be saved in your browser storage.&nbsp;
+        </p>
+      </v-card-title>
+      <v-card-text>
+        <v-text-field
+          :value="isValidAccount ? account.address.pretty() : ''"
+          class="ma-0 pa-0 monospaced"
+          label="Address"
+          readonly
+        />
+
+        <v-text-field
+          :value="isValidAccount ? account.publicKey : ''"
+          class="ma-0 pa-0 monospaced"
+          label="Public Key"
+          readonly
+        />
+
+        <v-text-field
+          v-model="node"
+          class="ma-0 pa-0"
+          label="NEM2 node URL"
+        >
+          {{ node }}
+        </v-text-field>
+
+        <v-text-field
+          v-model="walletName"
+          class="ma-0 pa-0"
+          label="Wallet name"
+        />
+
+        <v-text-field
+          v-model.lazy="privateKey"
+          class="ma-0 pa-0"
+          label="Private Key (64 char)"
+        />
+      </v-card-text>
+      <v-card-actions>
+        <v-spacer />
+        <v-btn
+          flat
+          @click.stop="$emit('close')"
+        >
+          Close
+        </v-btn>
+        <v-btn
+          :disabled="node == '' || walletName == '' || !isValidAccount"
+          color="primary mx-0"
+          @click="save"
+        >
+          Create Wallet
+        </v-btn>
+      </v-card-actions>
+    </v-card>
   </v-dialog>
 </template>
 
 <script>
-import WalletImport from './WalletImport.vue';
+import { NetworkType, Account } from 'nem2-sdk';
+import store from '../../store/index';
 
 export default {
   name: 'WalletImportDialog',
-  components: {
-    WalletImport,
-  },
+  store,
   props: {
     visible: Boolean,
+  },
+  data() {
+    return {
+      node: '',
+      walletName: '',
+      account: {},
+      privateKey: '',
+      isValidAccount: false,
+    };
   },
   computed: {
     show: {
@@ -45,6 +113,43 @@ export default {
           this.$emit('close');
         }
       },
+    },
+  },
+  watch: {
+    privateKey() {
+      this.createFromPrivateKey(this.privateKey);
+    },
+  },
+  methods: {
+    save() {
+      const newWallet = {
+        name: this.walletName,
+        account: this.account,
+        node: this.node,
+      };
+      this.$store.dispatch('wallet/ADD_WALLET', newWallet);
+      this.node = '';
+      this.walletName = '';
+      this.privateKey = '';
+      this.account = {};
+    },
+    createFromPrivateKey() {
+      const key = this.privateKey.length === 64
+        ? this.privateKey.toUpperCase() : false;
+      if (key) {
+        try {
+          this.account = Account.createFromPrivateKey(
+            key,
+            NetworkType.MIJIN_TEST,
+          );
+          this.isValidAccount = true;
+        } catch (error) {
+          this.isValidAccount = false;
+        }
+      } else {
+        this.account = {};
+        this.isValidAccount = false;
+      }
     },
   },
 };
