+++ src/components/wallet/WoWalletCreationDialog.vue
@@ -20,20 +20,80 @@
     v-model="show"
     max-width="680px"
   >
-    <WoWalletCreation />
+    <v-card>
+      <v-card-title primary-title>
+        <h3 class="headline mb-3">
+          Lookup another address
+        </h3>
+        <p class="mb-0">
+          Need to check this address again in the future? Store this address as a watch-only wallet!
+        </p>
+      </v-card-title>
+      <v-card-text>
+        <v-text-field
+          v-model="address"
+          class="ma-0 pa-0"
+          label="address"
+          type="text"
+          required
+        />
+        <v-text-field
+          v-model="node"
+          class="ma-0 pa-0"
+          label="node"
+          type="text"
+          required
+        />
+        <v-text-field
+          v-model="name"
+          class="ma-0 pa-0"
+          label="name (leave blank to use the address as a name)"
+          type="text"
+        />
+        <v-switch
+          v-model="isToBeSaved"
+          label="store as watch-only wallet"
+        />
+      </v-card-text>
+      <v-card-actions>
+        <v-spacer />
+        <v-btn
+          flat
+          @click.stop="$emit('close')"
+        >
+          close
+        </v-btn>
+        <v-btn
+          flat
+          :disabled="disabledValidation"
+          @click.stop="validateAddress"
+        >
+          OK
+        </v-btn>
+      </v-card-actions>
+    </v-card>
   </v-dialog>
 </template>
 
 <script>
-import WoWalletCreation from './WoWalletCreation.vue';
+import { Address } from 'nem2-sdk';
+import store from '../../store/index';
 
 export default {
   name: 'WoWalletCreationDialog',
-  components: {
-    WoWalletCreation,
-  },
+  store,
   props: {
     visible: Boolean,
+  },
+  data() {
+    return {
+      address: '',
+      validAddress: false,
+      name: '',
+      node: this.$store.getters['wallet/GET_ACTIVE_WALLET'].node,
+      disabledValidation: true,
+      isToBeSaved: true,
+    };
   },
   computed: {
     show: {
@@ -45,6 +105,48 @@ export default {
           this.$emit('close');
         }
       },
+    },
+  },
+  watch: {
+    address: {
+      handler(e) {
+        try {
+          const address = Address.createFromRawAddress(e);
+          this.validAddress = address;
+          this.disabledValidation = false;
+        } catch (error) {
+          this.validAddress = false;
+          this.disabledValidation = true;
+        }
+      },
+    },
+    name: {
+      handler(e) {
+        this.disabledValidation = !(typeof e === 'string');
+      },
+    },
+    node: {
+      handler(e) {
+        this.disabledValidation = !(typeof e === 'string');
+      },
+    },
+    isToBeSaved: {
+      handler(e) {
+        this.disabledValidation = !(typeof e === 'boolean');
+      },
+    },
+  },
+  methods: {
+    validateAddress() {
+      if (this.disabledValidation || !this.validAddress) return;
+      this.$store.dispatch('wallet/ADD_WATCH_ONLY_WALLET',
+        {
+          address: this.validAddress,
+          name: this.name === '' ? this.validAddress.pretty() : this.name,
+          node: this.node,
+          isToBeSaved: this.isToBeSaved,
+        });
+      this.$emit('close');
     },
   },
 };
